Verify source map support, inline main.ts
